trying out modified gzip module
